# [RFC] many: add simple performance measure tool mostly for firstboot