--- a/overlord/devicestate/firstboot.go
+++ b/overlord/devicestate/firstboot.go
@@ -31,6 +31,7 @@ import (
 	"github.com/snapcore/snapd/asserts/snapasserts"
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/i18n"
+	"github.com/snapcore/snapd/measure"
 	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/auth"
@@ -43,6 +44,9 @@ import (
 var errNothingToDo = errors.New("nothing to do")
 
 func installSeedSnap(st *state.State, sn *snap.SeedSnap, flags snapstate.Flags) (*state.TaskSet, *snap.Info, error) {
+	m := measure.New(fmt.Sprintf("installing for %s", sn.File))
+	defer m.Done()
+
 	if sn.Classic {
 		flags.Classic = true
 	}
@@ -80,6 +84,9 @@ func trivialSeeding(st *state.State, markSeeded *state.Task) []*state.TaskSet {
 }
 
 func populateStateFromSeedImpl(st *state.State) ([]*state.TaskSet, error) {
+	m := measure.New("populateStateFromSeed")
+	defer m.Done()
+
 	// check that the state is empty
 	var seeded bool
 	err := st.Get("seeded", &seeded)